Update tests for generic and custom blocks
